ordernum 추가 및 long type으로 모두 변경

+++ BeetMall/src/main/java/com/beetmall/sshj/admin/vo/Admin_WishVO.java
@@ -68,7 +68,7 @@ public class Admin_WishVO {
 	
 	//리뷰 VO
 	private int reviewnum;
-	private int ordernum;
+	private long ordernum;
 	private String reviewcontent;
 	private String reviewwritedate;
 	private int reviewrecommend;
@@ -332,10 +332,10 @@ public class Admin_WishVO {
 	public void setReviewnum(int reviewnum) {
 		this.reviewnum = reviewnum;
 	}
-	public int getOrdernum() {
+	public long getOrdernum() {
 		return ordernum;
 	}
-	public void setOrdernum(int ordernum) {
+	public void setOrdernum(long ordernum) {
 		this.ordernum = ordernum;
 	}
 	public String getReviewcontent() {
