+++ BeetMall/src/main/java/com/beetmall/sshj/custom/vo/MyBuyListVO.java
@@ -2,7 +2,7 @@ package com.beetmall.sshj.custom.vo;
 
 public class MyBuyListVO {
 	private String orderdate;
-	private int ordernum;
+	private long ordernum;
 	private String productname;
 	private String thumbimg;
 	private String productcontent;
@@ -25,10 +25,10 @@ public class MyBuyListVO {
 	public void setOrderdate(String orderdate) {
 		this.orderdate = orderdate;
 	}
-	public int getOrdernum() {
+	public long getOrdernum() {
 		return ordernum;
 	}
-	public void setOrdernum(int ordernum) {
+	public void setOrdernum(long ordernum) {
 		this.ordernum = ordernum;
 	}
 	public String getProductname() {
